+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## [0.0.9] - 2025-07-12
 ### Added
 - Permanent sidebar panel for device management.
+### Fixed
+- Use async static path registration to avoid runtime warnings.
 
 ## [0.0.8] - 2025-07-11
 ### Removed